Add files via upload
There are a couple of bugs that need fixing.  Updating the database is currently broken, it will crash and then require a couple of restarts until it works properly again.  Also, deleting deck seems to leave behind and empty cell.  Beyond bug fixes, the next step is to have the app download and present images of the cards in the detail view.